requirements.txt
phidata
youtube_transcript_api
groq
streamlit
python-dotenv
